fixed file structure
