+++ src/pages/Shop/CartSummary/SimilarProducts.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SimilarProducts = () => {
+  return <div>this is my smilar products</div>;
+};
+
+export default SimilarProducts;
